Implement an alternative service structure
This seperates the concerns of data retrieval (service/data/), model
representation and verification (service/domain/) and parameter parsing/view
model building (service/).

Hoping this will lead to reusability and readability. These are subjective to an
extent though so happy to take a steer.
